# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "忆述光华 API",
    "description": "个人记忆整理与回顾后端。功能：认证（微信/手机号/JWT）、内容入库（照片/文字/语音）、四层事件聚合、描述性搜索（RAG）、同步（LWW）、AI 分类/画像/护栏。契约对齐《忆述光华_Sprint1规划.md》S1-02。",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/wechat": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Wechat Login",
        "description": "微信登录：code 换 unionid → 建立/获取用户 → 签发 token 对（真实 DB）",
        "operationId": "wechat_login_api_v1_auth_wechat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TokenPair_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/phone": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Phone Login",
        "description": "手机号验证码登录（备用通道，真实校验 sms_codes；验证码哈希存储防 DB 泄漏）",
        "operationId": "phone_login_api_v1_auth_phone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TokenPair_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/sms/send": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send Sms",
        "description": "发送短信验证码（真实入库，6 位 + 5 分钟有效期 + 防刷限流 + 每日上限）",
        "operationId": "send_sms_api_v1_auth_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendSmsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "description": "refresh token 轮换（AUTH-005：旧 refresh 失效；devices 表可吊销 AUTH-006）",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TokenPair_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/classify": {
      "post": {
        "tags": [
          "classify"
        ],
        "summary": "Classify Text",
        "description": "文字碎片分类：入队异步执行（P2-01 推理移 worker），返回 job_id 供轮询",
        "operationId": "classify_text_api_v1_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/classify/jobs/{job_id}": {
      "get": {
        "tags": [
          "classify"
        ],
        "summary": "Classify Job Status",
        "description": "查询分类任务状态与结果（轮询；finished 时携带 result）",
        "operationId": "classify_job_status_api_v1_classify_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ClassifyJobOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/asr/transcribe": {
      "post": {
        "tags": [
          "asr"
        ],
        "summary": "Transcribe Audio",
        "description": "语音转写（API-004 主链路：上传 → 双通道转写 → 护栏审核）\n\n护栏（B5b）：转写结果先过 dashscope.moderate 再返回；\n拦截时 text 仍返回（供端侧展示），但 passed=false 提示不可下发。\n情绪（B5-c）：SenseVoice 通道产出 emotion，供情绪关怀分层触发。",
        "operationId": "transcribe_audio_api_v1_asr_transcribe_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "preferred",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(auto|funasr|sensevoice|mock)$",
              "default": "auto",
              "title": "Preferred"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_audio_api_v1_asr_transcribe_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AsrTranscribeResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/guard/check": {
      "post": {
        "tags": [
          "guard"
        ],
        "summary": "Guard Check",
        "description": "内容安全护栏（B5b）：发布/入库前文本审核，fail-safe 默认拦截（决策 #12）",
        "operationId": "guard_check_api_v1_guard_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_GuardCheckResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_AsrTranscribeResponse_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AsrTranscribeResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[AsrTranscribeResponse]"
      },
      "ApiResponse_ClassifyJobOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassifyJobOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[ClassifyJobOut]"
      },
...
      "ApiResponse_GuardCheckResponse_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GuardCheckResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[GuardCheckResponse]"
      },
...
      "ApiResponse_TokenPair_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TokenPair"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[TokenPair]"
      },
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "OK"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "default": ""
          }
        },
        "type": "object",
        "title": "ApiResponse[dict]"
      },
...
      "AsrTranscribeResponse": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "转写文本"
          },
          "channel": {
            "type": "string",
            "enum": [
              "funasr",
              "sensevoice",
              "mock"
            ],
            "title": "Channel",
            "description": "实际使用通道"
          },
          "emotion": {
            "type": "string",
            "title": "Emotion",
            "description": "声学情绪标签（SenseVoice 通道产出）",
            "default": "平静"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "description": "识别置信度",
            "default": 0.0
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Ms",
            "description": "音频时长（ms，wav 头解析）",
            "default": 0
          },
          "mock": {
            "type": "boolean",
            "title": "Mock",
            "description": "是否 mock 兜底",
            "default": false
          },
          "guardrail": {
            "$ref": "#/components/schemas/GuardrailVerdict",
            "description": "护栏审核结论（B5b）"
          }
        },
        "type": "object",
        "required": [
          "text",
          "channel",
          "guardrail"
        ],
        "title": "AsrTranscribeResponse",
        "description": "转写结果（mock 与真实同构）\n\n- channel: funasr（语义）/ sensevoice（声学情绪）/ mock（未配 key 兜底）\n- emotion: 声学情绪（开心/难过/生气/惊讶/恐惧/厌恶/平静），B5-c 情绪关怀分层触发依据\n- guardrail: 内容安全审核结论（fail-safe：真实模式下不可用默认拦截）"
      },
...
      "Body_transcribe_audio_api_v1_asr_transcribe_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "wav 音频（16kHz 16bit 单声道，≤8MB）"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_transcribe_audio_api_v1_asr_transcribe_post"
      },
...
      "ClassifyJobOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClassifyResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "ClassifyJobOut"
      },
      "ClassifyRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "description": "文字碎片内容"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ClassifyRequest"
      },
...
      "GuardCheckRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "description": "待审核文本（发布/入库前调用）"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "GuardCheckRequest"
      },
      "GuardCheckResponse": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "passed"
        ],
        "title": "GuardCheckResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PhoneLoginRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^1\\d{10}$",
            "title": "Phone",
            "description": "中国大陆手机号"
          },
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Code",
            "description": "短信验证码"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "PhoneLoginRequest",
        "description": "手机号验证码登录（备用通道）"
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "description": "refresh token 换新 access（轮换，旧 refresh 失效，AUTH-005）"
      },
...
      "SendSmsRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^1\\d{10}$",
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "SendSmsRequest",
        "description": "发送验证码（防刷：限流 + 有效期，测试清单 AUTH-003）"
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "access_expires_in": {
            "type": "integer",
            "title": "Access Expires In",
            "default": 7200
          },
          "refresh_expires_in": {
            "type": "integer",
            "title": "Refresh Expires In",
            "default": 2592000
          },
          "user": {
            "$ref": "#/components/schemas/UserBrief"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ],
        "title": "TokenPair"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WechatLoginRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "微信登录 code（wx.login 获取）"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id",
            "description": "客户端设备唯一 ID"
          }
        },
        "type": "object",
        "required": [
          "code",
          "device_id"
        ],
        "title": "WechatLoginRequest",
        "description": "APP 微信登录：code 换 openid/unionid（服务端调微信接口）"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}